Use the backdrop filter's clip for synthetic effects and layers
Previously for a synthetic effect and the associated synthesized mask
layer, we used the clip that introduce the synthetic effect as their
clip state, causing incorrect clipping of backdrop filters.

Now use the clip of the backdrop filter for synthetic effects and
layers to avoid the problem.

ClipPaintPropertyNode rounded_clip
  ClipPaintPropertyNode rect_clip

EffectPaintPropertyNode backdrop_filter clip=rect_clip

The cc effect tree is like
        mask_isolation (backdrop filter moved here)
          /       \
   mask_effect  node_for_backdrop_filter

When we create the cc node for backdrop_filter, we found that there is
an ancestor clip (rounded_clip) needs a synthetic effect and synthesized
mask, so create mask_isolation and mask_effect nodes and a mask clip
layer. To let the backdrop filter access the correct backdrop, we move
the backdrop filter up to mask_isolation, requiring that mask_isolation
to use the backdrop_filter's clip (rect_clip) instead of the clip
(rounded_clip) that introduces the synthetic effect, to ensure the
backdrop filter is correctly clipped. This in turn requires that
mask_effect and the mask layer also use rect_clip as their clip state to
ensure proper clip hierarchy.

Bug: 1077137
Change-Id: Ib9c3cc7e8db5b1b261b3a758f2fb6680e0f5ab1f